+++ api/src/main/java/org/allaymc/api/block/component/BlockBaseComponent.java
@@ -6,22 +6,15 @@ import org.allaymc.api.block.dto.Block;
 import org.allaymc.api.block.dto.PlayerInteractInfo;
 import org.allaymc.api.block.type.BlockState;
 import org.allaymc.api.block.type.BlockType;
-import org.allaymc.api.container.FullContainerType;
 import org.allaymc.api.entity.Entity;
-import org.allaymc.api.entity.component.EntityContainerHolderComponent;
-import org.allaymc.api.entity.component.EntityPhysicsComponent;
-import org.allaymc.api.entity.effect.EffectTypes;
 import org.allaymc.api.entity.interfaces.EntityProjectile;
 import org.allaymc.api.item.ItemStack;
-import org.allaymc.api.item.enchantment.type.EnchantmentTypes;
 import org.allaymc.api.world.Dimension;
 import org.jetbrains.annotations.ApiStatus;
 import org.joml.Vector3dc;
 import org.joml.Vector3ic;
 
 import java.util.Set;
-
-import static org.allaymc.api.item.ItemHelper.isSword;
 
 /**
  * @author daoge_cmd
@@ -313,74 +306,7 @@ public interface BlockBaseComponent extends BlockComponent {
      * @return the time in seconds required to break the block
      * @see <a href="https://minecraft.wiki/w/Breaking#Calculation">Breaking Calculation</a>
      */
-    default double calculateBreakTime(BlockState blockState, ItemStack usedItem, Entity entity) {
-        checkBlockType(blockState);
-        if (usedItem.canInstantBreak(blockState)) {
-            return 0;
-        }
-
-        var blockHardness = blockState.getBlockStateData().hardness();
-        if (blockHardness == -1) {
-            return Integer.MAX_VALUE;
-        }
-
-        var isCorrectTool = usedItem.isCorrectToolFor(blockState);
-        var requiresCorrectToolForDrops = blockState.getBlockStateData().requiresCorrectToolForDrops();
-
-        var baseTime = ((isCorrectTool || !requiresCorrectToolForDrops) ? 1.5d : 5d) * blockHardness;
-        var speed = 1d / baseTime;
-
-        double efficiency = 1d;
-        if (isCorrectTool) {
-            // Tool level (wooden, stone, iron, etc...) bonus
-            efficiency = usedItem.getBreakTimeBonus(blockState);
-            // Tool efficiency enchantment bonus
-            efficiency += speedBonusByEfficiency(usedItem.getEnchantmentLevel(EnchantmentTypes.EFFICIENCY));
-        }
-
-        if (isSword(usedItem.getItemType())) { // Special case
-            efficiency *= 1.5d;
-        }
-
-        speed *= efficiency;
-
-        if (entity != null) {
-            if (entity.hasEffect(EffectTypes.HASTE) || entity.hasEffect(EffectTypes.CONDUIT_POWER)) {
-                var level = Math.max(entity.getEffectLevel(EffectTypes.HASTE), entity.getEffectLevel(EffectTypes.CONDUIT_POWER));
-                speed *= (0.2d * level + 1) * Math.pow(1.2d, level);
-            }
-
-            // Entity mining fatigue effect negative bonus
-            if (entity.hasEffect(EffectTypes.MINING_FATIGUE)) {
-                // speedMultiplier *= 0.3 ^ miningFatigueLevel
-                // damage *= 0.7 ^ miningFatigueLevel
-                // 0.3 + 0.7 = 0.21 ^ miningFatigueLevel
-                speed *= Math.pow(0.21d, entity.getEffectLevel(EffectTypes.MINING_FATIGUE));
-            }
-
-            var hasAquaAffinity = false;
-            if (entity instanceof EntityContainerHolderComponent containerHolder) {
-                if (containerHolder.hasContainer(FullContainerType.ARMOR)) {
-                    hasAquaAffinity = containerHolder
-                            .getContainer(FullContainerType.ARMOR)
-                            .getHelmet()
-                            .hasEnchantment(EnchantmentTypes.AQUA_AFFINITY);
-                }
-            }
-
-            // In water but no underwater speed mining effect
-            if (entity.isEyesInWater() && !hasAquaAffinity) {
-                speed /= 5d;
-            }
-
-            // In air
-            if (entity instanceof EntityPhysicsComponent physicsComponent && !physicsComponent.isOnGround()) {
-                speed /= 5d;
-            }
-        }
-
-        return Math.ceil(1d / speed * 20d) / 20d;
-    }
+    double calculateBreakTime(BlockState blockState, ItemStack usedItem, Entity entity);
 
     /**
      * Check if a position on the side of the block placed in the world at a specific position is
@@ -393,19 +319,5 @@ public interface BlockBaseComponent extends BlockComponent {
      */
     default boolean canLiquidFlowIntoSide(BlockState blockState, BlockFace blockFace) {
         return true;
-    }
-
-    private double speedBonusByEfficiency(int efficiencyLevel) {
-        if (efficiencyLevel == 0) {
-            return 0;
-        }
-
-        return (efficiencyLevel * efficiencyLevel) + 1;
-    }
-
-    private void checkBlockType(BlockState blockState) {
-        if (blockState.getBlockType() != getBlockType()) {
-            throw new IllegalArgumentException("Block type is not match! Expected: " + getBlockType().getIdentifier() + ", actual: " + blockState.getBlockType().getIdentifier());
-        }
     }
 }