+++ api/src/main/java/org/allaymc/api/item/enchantment/EnchantmentTypes.java
@@ -1,0 +1,49 @@
+package org.allaymc.api.item.enchantment;
+
+/**
+ * @author IWareQ | daoge_cmd
+ */
+@SuppressWarnings("unused")
+public final class EnchantmentTypes {
+    public static EnchantmentType AQUA_AFFINITY;
+    public static EnchantmentType BANE_OF_ARTHROPODS;
+    public static EnchantmentType BINDING;
+    public static EnchantmentType BLAST_PROTECTION;
+    public static EnchantmentType BREACH;
+    public static EnchantmentType CHANNELING;
+    public static EnchantmentType DENSITY;
+    public static EnchantmentType DEPTH_STRIDER;
+    public static EnchantmentType EFFICIENCY;
+    public static EnchantmentType FEATHER_FALLING;
+    public static EnchantmentType FIRE_ASPECT;
+    public static EnchantmentType FIRE_PROTECTION;
+    public static EnchantmentType FLAME;
+    public static EnchantmentType FORTUNE;
+    public static EnchantmentType FROST_WALKER;
+    public static EnchantmentType IMPALING;
+    public static EnchantmentType INFINITY;
+    public static EnchantmentType KNOCKBACK;
+    public static EnchantmentType LOOTING;
+    public static EnchantmentType LOYALTY;
+    public static EnchantmentType LUCK_OF_THE_SEA;
+    public static EnchantmentType LURE;
+    public static EnchantmentType MENDING;
+    public static EnchantmentType MULTISHOT;
+    public static EnchantmentType PIERCING;
+    public static EnchantmentType POWER;
+    public static EnchantmentType PROJECTILE_PROTECTION;
+    public static EnchantmentType PROTECTION;
+    public static EnchantmentType PUNCH;
+    public static EnchantmentType QUICK_CHARGE;
+    public static EnchantmentType RESPIRATION;
+    public static EnchantmentType RIPTIDE;
+    public static EnchantmentType SHARPNESS;
+    public static EnchantmentType SILK_TOUCH;
+    public static EnchantmentType SMITE;
+    public static EnchantmentType SOUL_SPEED;
+    public static EnchantmentType SWIFT_SNEAK;
+    public static EnchantmentType THORNS;
+    public static EnchantmentType UNBREAKING;
+    public static EnchantmentType VANISHING;
+    public static EnchantmentType WIND_BURST;
+}